+++ src/common.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use crate::debug_dll::{get_string_from_memory_bounded, save_string_to_memory};
+use crate::util::{get_string_from_memory_bounded, save_string_to_memory};
 
 #[derive(Debug)]
 #[repr(C)]
