+++ src/lib.rs
@@ -64,6 +64,9 @@ mod version;
 /// OpenZT mod structs
 mod mods;
 
+/// Utility functions for working with the game's memory, including reading and writing memory, and patching the game's assembly.
+mod util;
+
 #[cfg(target_os = "windows")]
 use windows::Win32::System::SystemServices::{DLL_PROCESS_ATTACH, DLL_PROCESS_DETACH, DLL_THREAD_ATTACH, DLL_THREAD_DETACH};
 
@@ -121,11 +124,6 @@ extern "system" fn DllMain(module: u8, reason: u32, _reserved: u8) -> i32 {
                 add_to_command_register("get_setting".to_owned(), command_get_setting);
                 add_to_command_register("set_setting".to_owned(), command_set_setting);
             }
-            if cfg!(feature = "bf_registry") {
-                use crate::bfregistry;
-                info!("Feature 'bf_registry' enabled");
-                bfregistry::init();
-            }
 
             if cfg!(feature = "capture_ztlog") {
                 use crate::capture_ztlog;
@@ -159,79 +157,41 @@ extern "system" fn DllMain(module: u8, reason: u32, _reserved: u8) -> i32 {
 
 // TODO: Make sure there isn't anything useful here before removing
 // #[no_mangle]
-// extern "C" fn dll_ini_debug_log() {
-//     debug_dll::log_debug_ini_memory_values();
-// }
-
-// fn load_debug_settings_from_ini() {
-//     debug_dll::debug_logger("load_debug_settings_from_ini");
-//     debug_dll::log_exe_location_memory_value();
-//     debug_dll::log_debug_ini_memory_values();
-//     let mut base_path = debug_dll::get_base_path();
-//     base_path.push("zoo.ini");
-//     let debug_settings = load_ini::load_debug_settings(base_path.as_path());
-//     debug_dll::debug_logger("Saving debug ini settings");
-//     debug_dll::save_debug_settings(debug_settings);
-//     debug_dll::log_debug_ini_memory_values();
-// }
-
-// #[no_mangle]
-// pub fn patch_load_debug_ini_call() {
-//     debug_dll::debug_logger(&format!("load_debug_settings_from_ini {:p}", load_debug_settings_from_ini as *const ()));
-//     debug_dll::debug_logger(&format!("load_debug_settings_from_ini (u32) {}", load_debug_settings_from_ini as u32));
-//     debug_dll::get_code_from_memory(debug_dll::DEBUG_INI_LOAD_CALL_ADDRESS, 0x10);
-//     debug_dll::patch_call(debug_dll::DEBUG_INI_LOAD_CALL_ADDRESS, load_debug_settings_from_ini as u32);
-// }
-
-// #[no_mangle]
-// extern "C" fn patch_load_int_from_ini_call() {
-//     debug_dll::debug_logger(&format!("load_int_from_ini {:p}", load_int_from_ini as *const ()));
-//     debug_dll::patch_calls(debug_dll::LOAD_INT_FROM_INI_ADDRESS_ARRAY_SUBSET.to_vec(), load_int_from_ini as u32);
-//     debug_dll::patch_nops_series(debug_dll::LOAD_INT_FROM_INI_ADDRESS_ARRAY_SUBSET_NOP.to_vec());
-// }
-
-// #[no_mangle]
-// extern "C" fn patch_load_value_from_ini_call() {
-//     debug_dll::debug_logger(&format!("load_value_from_ini {:p}", load_value_from_ini as *const ()));
-//     debug_dll::patch_calls(debug_dll::LOAD_VALUE_FROM_INI_ADDRESS_ARRAY.to_vec(), load_value_from_ini as u32);
-// }
-
-// #[no_mangle]
 // extern "cdecl" fn load_int_from_ini(section_address: &u32, header_address: &u32, default: i32) -> u32 {
-//     debug_dll::debug_logger(&format!(
+//     util::debug_logger(&format!(
 //         "load_int_from_ini {:p} {:p} default: {}",
 //         *section_address as *const (), *header_address as *const (), default
 //     ));
-//     let section = debug_dll::get_string_from_memory(*section_address);
-//     let header = debug_dll::get_string_from_memory(*header_address);
+//     let section = util::get_string_from_memory(*section_address);
+//     let header = util::get_string_from_memory(*header_address);
 //     let mut zoo_ini = Ini::new();
 //     zoo_ini.load(get_ini_path()).unwrap();
 //     let result = load_ini::load_int_with_default(&zoo_ini, &section, &header, default) as u32;
-//     debug_dll::debug_logger(&format!("load_int_from_ini {} {} result: {}", section, header, result));
+//     util::debug_logger(&format!("load_int_from_ini {} {} result: {}", section, header, result));
 //     result
 // }
 
 // #[no_mangle]
 // extern "cdecl" fn load_value_from_ini<'a>(result_address: &'a u32, section_address: &u32, header_address: &u32, default_address: &u32) -> &'a u32 {
-//     debug_dll::debug_logger(&format!(
+//     util::debug_logger(&format!(
 //         "load_value_from_ini {:p} {:p} default: {:p}",
 //         *section_address as *const (), *header_address as *const (), *default_address as *const ()
 //     ));
-//     let section = debug_dll::get_string_from_memory(*section_address);
-//     let header = debug_dll::get_string_from_memory(*header_address);
-//     let default = debug_dll::get_string_from_memory(*default_address);
+//     let section = util::get_string_from_memory(*section_address);
+//     let header = util::get_string_from_memory(*header_address);
+//     let default = util::get_string_from_memory(*default_address);
 //     let mut zoo_ini = Ini::new();
 //     zoo_ini.load(get_ini_path()).unwrap();
 //     let result = load_ini::load_string_with_default(&zoo_ini, &section, &header, &default);
 
-//     debug_dll::debug_logger(&format!("load_value_from_ini {} {} result: {}", section, header, result));
-//     debug_dll::debug_logger(&format!("encoding string at address: {:p}", *result_address as *const ()));
-//     debug_dll::save_string_to_memory(*result_address, &result);
+//     util::debug_logger(&format!("load_value_from_ini {} {} result: {}", section, header, result));
+//     util::debug_logger(&format!("encoding string at address: {:p}", *result_address as *const ()));
+//     util::save_string_to_memory(*result_address, &result);
 //     result_address
 // }
 
 // fn get_ini_path() -> String {
-//     let mut base_path = debug_dll::get_base_path();
+//     let mut base_path = util::get_base_path();
 //     base_path.push("zoo.ini");
 //     base_path.to_str().unwrap().to_string()
 // }
