+++ src/version.rs
@@ -9,7 +9,7 @@ pub fn init() {
 
 #[hook_module("zoo.exe")]
 mod bf_version_info {
-    use crate::debug_dll::{get_from_memory, get_string_from_memory, save_string_to_memory, save_to_memory};
+    use crate::util::{get_from_memory, get_string_from_memory, save_string_to_memory, save_to_memory};
 
     #[hook(unsafe extern "cdecl" BFVersionInfo_GetVersionStringHook, offset = 0x000bdfd4)]
     fn bf_version_info_get_version_string_hook(param_1: u32, param_2: u32, param_3: u32) -> u32 {
