+++ src/ztgamemgr.rs
@@ -1,6 +1,6 @@
 use tracing::info;
 
-use crate::{add_to_command_register, console::CommandError, debug_dll::get_from_memory};
+use crate::{add_to_command_register, console::CommandError, util::get_from_memory};
 
 const GLOBAL_ZTGAMEMGR_ADDRESS: u32 = 0x00638048;
