+++ src/ztui.rs
@@ -5,7 +5,7 @@ use tracing::info;
 use crate::{
     common::ZTString,
     console::{add_to_command_register, CommandError},
-    debug_dll::{get_from_memory, get_string_from_memory_bounded},
+    util::{get_from_memory, get_string_from_memory_bounded},
     ztworldmgr::read_zt_entity_from_memory,
 };
